Lots of changes to layout of project.  Disables the editing of providers temporarily
Adding or changing providers is broken because it will not convert them
to array.  This will also limit the number of results to 100 at a time.
Some sort of pagination can be added later.

This also changes how the proxy works to allow for local api testing
instead of just forwarding all requests to server